pyproject.toml
[project]
name = "llm-council"
version = "0.1.0"
description = "Medical QA Benchmark - Multi-Agent Debate"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "datasets>=3.0.0",
    "scikit-learn>=1.5.0",
]
